requirements.txt
langchain==1.2.3
langchain_community==0.4.1
langchain_core==1.2.7
langchain_groq==1.1.1
langchain_huggingface==1.2.0
langchain_pinecone==0.2.13
langchain_text_splitters==1.1.0
python-dotenv==1.2.1
streamlit==1.52.2
